+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## dev ##
 
 ### enhancements
-  * Array-like attributes. (by [@lest](https://github.com/lest))
+  * SimpleForm support for multiple attributes. (by [@nashby](https://github.com/nashby))
+  * Array-like multiple attributes. (by [@lest](https://github.com/lest))
 
 ## 0.4.0 (September 6, 2012) ##